fixing bug in tests
