+++ load_data.txt
@@ -4,3 +4,33 @@ GODSON UKPERE STAFF
 JOAN NGATIA FELLOW Y
 HELLEN May STAFF
 BRIAN MOSIGISI FELLOW Y
+ALEX MWALEH FELLOW
+NAD ENEGESI STAFF
+JOSHUA MWANIKI STAFF
+JACKSON SAYA STAFF
+NATE MARTIN FELLOW Y
+EUGENE MUTAI FELLOW Y
+JOHN KARIUKI FELLOW Y
+SUADA HAJI FELLOW Y
+AMOS OMOSH FELLOW Y
+ANDREW ONYANGO FELLOW
+PENINAH WANJIRU FELLOW Y
+EDWIN KUPARA FELL
+ESTON KARUMBI FELLOW
+CHRIS GANGA FELLOW Y
+STANLEY NDAGI FELLOW Y
+GITHINJI GIKERA FELLOW Y
+STACY ACHUNGO FELLOW Y
+GETTIE NYENYESHI FELLOW Y
+JEREMY KITHOME FELLOW Y
+LEWIS MAINA FELLOW Y
+COLLINS MUTEMBEI FELLOW
+MAHAD WALUSIMBI FELLOW Y
+THOMAS NYAMBATI FELLOW Y
+PEGGIE WANJIRU FELLOW Y
+PUTA GUTTA STAFF
+MAYWEATHER FLOYD STAFF
+JOHN MAYER FELLOW N
+JOHN FRUSCIANTE STAFF
+HAILEY WILLIAMS FELLOW N
+NICOLAI CAGE STAFF
